ChainServer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentorchestrator"
version = "0.1.0"
description = "A DAG-based chain orchestration framework for AI/ML pipelines"
readme = "agentorchestrator/README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["ai", "ml", "chains", "dag", "orchestration", "llm", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

# Core dependencies - minimal footprint
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# LLM summarization support
summarization = [
    "tiktoken>=0.5.0",
    "langchain-core>=0.1.0",
    "langchain-text-splitters>=0.0.1",
]
# OpenAI provider
openai = [
    "langchain-openai>=0.0.5",
]
# Anthropic provider
anthropic = [
    "langchain-anthropic>=0.1.0",
]
# Observability (tracing, structured logging)
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "structlog>=23.0.0",
]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
# All optional features
all = [
    "agentorchestrator[summarization,openai,anthropic,observability]",
]

[project.scripts]
agentorchestrator = "agentorchestrator.cli:main"
ao = "agentorchestrator.cli:main"

[project.urls]
Repository = "https://github.com/your-org/agentorchestrator"

# ============================================================
# BUILD
# ============================================================
[tool.hatch.build.targets.wheel]
packages = ["agentorchestrator"]

[tool.hatch.build.targets.sdist]
include = ["/agentorchestrator", "/tests"]

# ============================================================
# RUFF - Linting & Formatting
# ============================================================
[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = [
    "E501",   # line too long
    "E402",   # module level import not at top (needed for conditional imports)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]
"agentorchestrator/examples/*" = ["F841", "W293"]
"agentorchestrator/core/*" = ["F841"]

[tool.ruff.lint.isort]
known-first-party = ["agentorchestrator"]

# ============================================================
# PYTEST
# ============================================================
[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["agentorchestrator/tests"]
python_files = ["test_*.py"]
addopts = ["-ra", "-q", "--strict-markers"]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
]
filterwarnings = ["ignore::DeprecationWarning"]

# ============================================================
# COVERAGE
# ============================================================
[tool.coverage.run]
source = ["agentorchestrator"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 70
